+++ server/src/main/java/com/groop/server/dto/TaskDTO.java
@@ -1,6 +1,6 @@
 package com.groop.server.dto;
 
-import com.groop.server.domain.TaskStatus;
+import com.groop.server.model.TaskStatus;
 
 /**
  * @author joandy alejo garcia
