+++ server/src/main/java/com/groop/server/model/User.java
@@ -1,4 +1,4 @@
-package com.groop.server.domain;
+package com.groop.server.model;
 import javax.persistence.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
