+++ server/src/main/java/com/groop/server/service/CommentService.java
@@ -1,7 +1,7 @@
 package com.groop.server.service;
 
-import com.groop.server.domain.Comment;
-import com.groop.server.domain.Task;
+import com.groop.server.repository.CommentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -9,8 +9,15 @@ import java.util.Optional;
 /**
  * @author joandy alejo garcia
  */
-public interface CommentService {
-    void deleteComment(Comment comment);
+@Service
+public class CommentService {
+    @Autowired
+    CommentRepository commentRepository;
+    public void deleteComment(com.groop.server.model.Comment comment) {
+        commentRepository.delete(comment);
+    }
 
-    Optional<Comment> findCommentById(Long comment_id);
+    public Optional<com.groop.server.model.Comment> findCommentById(Long comment_id) {
+        return commentRepository.findById(comment_id);
+    }
 }
